test using iam role

+++ spot/spot_manager.py
@@ -60,7 +60,8 @@ class SpotManager(object):
         # ADD UP THE CURRENT REQUESTED INSTANCES
         active = qb.filter(spot_requests, {"terms": {"status.code": RUNNING_STATUS_CODES | PENDING_STATUS_CODES}})
         for a in active:
-            Log.note("Active Spot Request: {{type}} @ {{price|round(decimal=4)}}", {
+            Log.note("Active Spot Request {{id}}: {{type}} @ {{price|round(decimal=4)}}", {
+                "id": a.id,
                 "type": a.launch_specification.instance_type,
                 "price": a.price
             })
@@ -78,7 +79,7 @@ class SpotManager(object):
         net_new_utility = utility_required - current_utility
 
         if net_new_utility <= 0:
-            net_new_utility += self.remove_instances(-net_new_utility)
+            net_new_utility = self.remove_instances(net_new_utility)
 
         if net_new_utility > 0:
             net_new_utility, remaining_budget = self.add_instances(net_new_utility, remaining_budget)
@@ -108,7 +109,7 @@ class SpotManager(object):
                 })
                 continue
 
-            num = Math.floor(net_new_utility / p.type.utility)
+            num = Math.round(net_new_utility / p.type.utility)
             if num == 1:
                 min_bid = max_bid
                 price_interval = 0
@@ -142,7 +143,7 @@ class SpotManager(object):
 
         return net_new_utility, remaining_budget
 
-    def remove_instances(self, utility_to_remove):
+    def remove_instances(self, net_new_utility):
         # FIND THE BIGGEST, MOST EXPENSIVE REQUESTS
         instances = self._get_managed_instances()
 
@@ -151,13 +152,13 @@ class SpotManager(object):
 
         instances = qb.sort(instances, [
             {"value": "markup.type.utility", "sort": -1},
-            {"value": "markup.estimated_value", "sort": -1}
+            {"value": "markup.estimated_value", "sort": 1}
         ])
 
         # FIND COMBO THAT WILL SHUTDOWN WHAT WE NEED EXACTLY, OR MORE
         remove_list = []
         for acceptable_error in range(0, 8):
-            remaining_utility = utility_to_remove
+            remaining_utility = -net_new_utility
             remove_list = DictList()
             for s in instances:
                 utility = coalesce(s.markup.type.utility, 0)
@@ -165,31 +166,36 @@ class SpotManager(object):
                     remove_list.append(s)
                     remaining_utility -= utility
             if remaining_utility <= 0:
+                net_new_utility = -remaining_utility
                 break
 
+        if not remove_list:
+            return net_new_utility
+
         # SEND SHUTDOWN TO EACH INSTANCE
+        Log.note("Shutdown {{instances}}", {"instances": remove_list.id})
         for i in remove_list:
             try:
                 self.instance_manager.teardown_instance(i)
             except Exception, e:
                 Log.warning("Teardown of {{id}} failed", {"id": i.id}, e)
 
-        remove_requests = remove_list.spot_instance_request_id
+        remove_spot_requests = remove_list.spot_instance_request_id
 
         # TERMINATE INSTANCES
         self.conn.terminate_instances(instance_ids=remove_list.id)
 
         # TERMINATE SPOT REQUESTS
-        self.conn.cancel_spot_instance_requests(request_ids=remove_requests)
+        self.conn.cancel_spot_instance_requests(request_ids=remove_spot_requests)
 
-        return -remaining_utility  # RETURN POSITIVE NUMBER IF TOOK AWAY TOO MUCH
+        return net_new_utility
 
     def _get_managed_instances(self):
         output =[]
         reservations = self.conn.get_all_instances()
         for res in reservations:
             for instance in res.instances:
-                if instance.tags.get('Name', '').startswith(self.settings.ec2.instance.name):
+                if instance.tags.get('Name', '').startswith(self.settings.ec2.instance.name) and instance._state.name == "running":
                     output.append(dictwrap(instance))
         return wrap(output)
 
@@ -212,7 +218,7 @@ class SpotManager(object):
                             self.instance_manager.setup_instance(i, p.type.utility)
                             i.add_tag("Name", self.settings.ec2.instance.name + " (running)")
                         except Exception, e:
-                            Log.warning("problem with setup of {{instance_id}}", {"instance_id": i.id})
+                            Log.warning("problem with setup of {{instance_id}}", {"instance_id": i.id}, e)
 
                 if self.done_spot_requests and not extra_refresh:  # NOTICE THE CHANGE IN done_spot_requests WE CAN EXPECT MORE PENDING REQUESTS
                     extra_refresh = self.done_spot_requests.is_go()
